# archive/backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PushSubscriptionKeys": {
        "properties": {
          "auth": {
            "title": "Auth",
            "type": "string"
          },
          "p256dh": {
            "title": "P256Dh",
            "type": "string"
          }
        },
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "PushSubscriptionKeys",
        "type": "object"
      },
      "PushSubscriptionRequest": {
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeys"
          }
        },
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI personal assistant for ADHD",
    "title": "Unspool",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/push/subscribe": {
      "post": {
        "operationId": "push_subscribe_api_push_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Push Subscribe Api Push Subscribe Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Subscribe"
      }
    },
...
  }
}